js/xpconnect/tests/mochitest/test_bug800864.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=800864
-->
<head>
  <title>Test for Bug 800864</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=800864">Mozilla Bug 800864</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

function checkThrows(f) {
  try {
    f();
    ok(false, "Didn't throw a security exception like we should");
  } catch(e) {
    ok(/denied|insecure/.exec(e), "Should throw security exception. Got: " + e);
  }
}

function go() {
  ifr = document.getElementById('ifr');
  win = ifr.contentWindow;
  loc = win.location;
  ifr.onload = check;
  win.location = 'http://test1.example.com';
}

function check() {
  checkThrows(function() { loc.toString(); });
  checkThrows(function() { loc.valueOf().toString(); });
  checkThrows(function() { loc.href; });
  checkThrows(function() { loc + ''; });
  SimpleTest.finish();
}

</script>
</pre>
</body>
<iframe id="ifr" onload="go()" src="file_empty.html"></iframe>
</html>
